Web Server/requirements.txt
fastapi
uvicorn[standard]
paho-mqtt
pymysql
python-dotenv
requests
